+++ Program.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.Configuration;
 using System.Diagnostics;
+using VectorIndexScenarioSuite.filtersearch;
 
 
 namespace VectorIndexScenarioSuite
@@ -64,8 +65,8 @@ namespace VectorIndexScenarioSuite
 
             switch (scenarios)
             {
-                case Scenarios.AmazonAutomotive:
-                    return new AmazonAutomotiveScenario(configurations);
+                case Scenarios.AutomotiveEcommerce:
+                    return new AutomotiveEcommerceScenario(configurations);
                 case Scenarios.MSMarcoEmbeddingOnly:
                     return new MSMarcoEmbeddingOnlyScenario(configurations);
                 case Scenarios.MSTuringEmbeddingOnly:
