--- dataset/automotive_label_download.ps1
@@ -1,1 +1,0 @@
-# working on getting the dataset for public use
